left notes for all db and server functions

+++ server/database/database.js
@@ -10,22 +10,24 @@ const con = mysql.createConnection({
 
 con.connect(function (err) {
     if (err) {
-        console.error(err);
+        //delete console log if functions correctly
+        //console.error(err);
         return;
     }
-    console.log('Connected to RDS MySQL database!');
+   // console.log('Connected to RDS MySQL database!');
 });
 
 
 
 //addNewUser will store the new profiles from the signUp component
 const addNewUser = (newUser, callback) => {
-    let queryString =  `INSERT into USER (Name, Username, Password, Email) VALUES 
+    let queryString =  `INSERT into USER (name, username, password, email) VALUES 
     ('${newUser.name}','${newUser.userName}', '${newUser.password}', '${newUser.email}')`
 
     con.query(queryString, (err, data) => {
         if(err) {
-            console.log('>>> could not add user', err);
+            //delete console log if functions correctly
+           // console.log('>>> could not add user', err);
         }
         callback(null, data);
     })
@@ -33,12 +35,13 @@ const addNewUser = (newUser, callback) => {
 
 //accessUser will retrieve both the Username and the Password from the User Table
 const accessUser = (profile, callback) => {
-    let queryString = `SELECT (Username, Password) 
-    FROM Users WHERE Username=${profile.userName}, Password=${profile.password}'}`
+    let queryString = `SELECT (userID) 
+    FROM Users WHERE username=${profile.userName}, password=${profile.password}'}`
 
     con.query(queryString, (err, data) => {
         if(err) {
-            console.log('>>> could not find user', err);
+            //delete console log if functions correctly
+            //console.log('>>> could not find user', err);
         }
         callback(null, data);
     })
@@ -46,11 +49,12 @@ const accessUser = (profile, callback) => {
 
 //removeUser deletes 
 const removeUser = (deleteProfile, callback) => {
-    let queryString = `DELETE FROM Users WHERE UsersID=${deleteProfile}`
+    let queryString = `DELETE FROM Users WHERE UsersID=${deleteProfile.UserID}`
 
     con.query(queryString, (err, data) => {
         if(err) {
-            console.log('>>> could not delete profile from Users')
+            //delete console log if functions correctly
+           // console.log('>>> could not delete profile from Users')
         }
         callback(null, data);
     })
@@ -61,49 +65,41 @@ const removeUser = (deleteProfile, callback) => {
 
 //needs to access user
 const addFoodToPantry = (newItem, callback) => {
-    let queryString = `INSERT into Foods ('${newItem.item_name}, ${newItem.expiration})
-    SELECT ${newItem.userName} FROM User, Users_userName,
-    WHERE Users_userName=${newItem.Users_userName}`
+      /*INSERT into Foods (item_name, expiration)
+      select food_item, expiration from foods where Users_UserID 
+      IN (select userID from USERs where username=${newItem.name})*/
+    let queryString = `INSERT into Foods (item_name, expiration) VALUES (${newItem.item_name}, ${newItem.expiration})
+  
+    SELECT ${newItem.item_name}, ${newItem.expiration} FROM Foods
+    WHERE Users_UserID IN (SELECT UserId FROM Users WHERE username=${newItem.name})`
 
     con.query(queryString, (err, data) => {
         if(err) {
-            console.log('>>> could not add food to pantry', err);
+            //delete console log if functions correctly
+            //console.log('>>> could not add food to pantry', err);
         }
         callback(null, data);
     })
 };
-
-
-
-//addExpirationDate will add an expiration date to a food item
-
-//needs to access user
-// const addExpirationDate = (newDate, callback) => {
-//     let queryString = `INSERT into EXP ('${newDate.expiration}')`
-
-//     con.query(queryString, (err, data) => {
-//         if(err) {
-//             console.log('>>> could not add expiration to food item', err);
-//         }
-//         callback(null, data);
-//     })
-// };
 
 
 //userData will show an existing user's info and pantry items
 const userData = (info, callback) => {
-    let queryString = `SELECT (Users.Name=${info.name}, Foods.Item=${info.item_name}, EXP.Expiration_Date=${info.expiration}) 
-    FROM Users, Foods, EXP`
+    let queryString = `SELECT * FROM Foods WHERE Users_UserID=${info}`
+    //possibility you may to access through Users_UserID=${info.id}
 
+//select food.item, expiration from foods where Users_UserID IN (select userID from USERs where username=${info.name})
     con.query(queryString, (err, data) => {
         if(err) {
-            console.log(`>>> could not the user's info`, err)
+            //delete console log if functions correctly
+           // console.log(`>>> could not the user's info`, err)
         }
         callback(null, data);
     })
 };
 
-
+//needs to pull from the users table, query from the foods table, with that user id
+//on the client side, 
 
 
 //removePantryItem will remove the item posted in your inventory
@@ -112,7 +108,8 @@ const removePantryItem = (removeItem, callback) => {
 
     con.query(queryString, (err, data) => {
         if(err) {
-            console.log('>>> could not delete the food item', err)
+            //delete console log if functions correctly
+            //console.log('>>> could not delete the food item', err)
         }
         callback(null, data);
     })
@@ -131,5 +128,5 @@ A FOREIGN KEY with cascade delete means that if a record in the parent table is 
 then the corresponding records in the child table will automatically be deleted.
 */
 
-module.exports = {con, addNewUser, accessUser, removeUser, addFoodToPantry, addExpirationDate, userData, removePantryItem,}
+module.exports = {con, addNewUser, accessUser, removeUser, addFoodToPantry, userData, removePantryItem,}
